Search Tool Modification and New Libraries
Libraries: Static: Dragtablejs and Panelsnap. I'm not using Panelsnap at
the moment. DragtableJS is for ordering the column on a table by
clicking on them.

Major changes: Added the interactive partial and full tables (horizontal
and vertical mode) to the Search Tool. Touched the .css files. Need to
touch them more.